Resolving conflicts and fixing code samples
